Move RenderTargetC to a separate file.
